requirements.txt
playwright==1.52.0
boto3
requests
flask
python-dotenv
